adding pagination via connector_cli
Tested locally but I'm unsure whether the paging and/or system works as intended. When I change from and count in the system config i.e:

"list": {
      "id_expression": "{{ id }}",
      "method": "GET",
      "next_page_link": "$ENV(base_url)/v2/customer?from=3&count=2",
      "next_page_termination_strategy": ["empty-result", "same-next-page-link"],
      "payload_property": "values",
      "url": "{{ properties.datatype }}"
    },

I get that the pipe processes a total of 9 entities. When I go into the collect pipe and enable preview, I see that I get 3 duplicates now which is seen during the run. I compared two of them via: https://jsondiff.com/ . They were semantically identical.

My question is, why does this happen and it is intended? I can't gather why that would be the case.